pyproject.toml
[project]
name = "cafe-badminton-uploader"
version = "0.1.0"
description = "Upload videos to Cafe Badminton channel"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clip",
    "freezegun>=1.5.5",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "imagehash>=4.3.2",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "temporalio>=1.21.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
